wip: fix import cycles
